+++ api/client/container/stats.go
@@ -18,16 +18,12 @@ import (
 	"github.com/docker/engine-api/types/events"
 	"github.com/docker/engine-api/types/filters"
 	"github.com/spf13/cobra"
-
-
 )
 
 type statsOptions struct {
 	all      bool
 	noStream bool
 	v	 bool
-
-
 	containers []string
 }
 
@@ -93,7 +89,6 @@ func runStats(dockerCli *client.DockerCli, opts *statsOptions) error {
 	waitFirst := &sync.WaitGroup{}
 
 	cStats := stats{}
-	vStats := vstats{}
 	// getContainerList simulates creation event for all previously existing
 	// containers (only used when calling `docker stats` without arguments).
 	getContainerList := func() {
@@ -163,92 +158,8 @@ func runStats(dockerCli *client.DockerCli, opts *statsOptions) error {
 			fmt.Println("Please provide container name(s)")
 			return nil
 		}
-				
-		for _,name:=range opts.containers{
-			s := &volumeStats{container: name}
-			if vStats.add_v(s){
-				//collects list of volumes for each container
-				s.CollectVol(ctx,dockerCli.Client())
-			}	
-		}
-		go func(){
-		for{	
-			vStats.mu.Lock()
-			for _,s:=range vStats.vs{
-				if s.err != nil{
-					fmt.Println(s.err)
-					time.Sleep(100 * time.Millisecond)
-					continue
-				}
-				//collects volume stats for each volume
-				s.CollectVolStats(ctx,dockerCli.Client())
-			}
-			vStats.mu.Unlock()
-			time.Sleep(300*time.Millisecond)
-			if opts.noStream{
-				break
-			}
-		}
-		}()//go
-		close(closeChan)
-		var errs []string
-		vStats.mu.Lock()
-		for _, c := range vStats.vs {
-			if c.err != nil {
-				errs = append(errs, fmt.Sprintf("%s: %v", c.container, c.err))
-			}
-		}
-		vStats.mu.Unlock()
-		if len(errs) > 0 {
-			return fmt.Errorf("%s", strings.Join(errs, ", "))
-		}
-		printHeader := func() {
-			if !opts.noStream{
-			  	fmt.Fprint(dockerCli.Out(), "\033[2J")
-				fmt.Fprint(dockerCli.Out(), "\033[H")
-			}
-		
-		}
-		for range time.Tick(500 * time.Millisecond) {
-			vStats.mu.Lock()
-			printHeader()
-			toRemove := []string{}
-			for _, s := range vStats.vs {
-				if err := s.DisplayVol(); err != nil && !opts.noStream{
-					logrus.Debugf("stats: got error for %s: %v", s.container, err)
-					if err == io.EOF {
-						toRemove = append(toRemove, s.container)
-					}
-				}
-			}
-		for _, name := range toRemove {
-			vStats.remove_v(name)
-		}
-		if len(vStats.vs) == 0{
-			vStats.mu.Unlock()
-			return nil
-		}
-		vStats.mu.Unlock()
-		if opts.noStream {
-			break
-		}
-		select {
-		case err, ok := <-closeChan:
-			if ok {
-				if err != nil {
-					// this is suppressing "unexpected EOF" in the cli when the
-					// daemon restarts so it shutdowns cleanly
-					if err == io.ErrUnexpectedEOF {
-						return nil
-					}
-					return err
-				}
-			}
-		default:
-			// just skip
-		}
-	}
-	return nil
+		RunvStats(ctx,dockerCli,opts.containers,opts.noStream,closeChan)	
+		return nil
 
 	}else{
 		// Artificially send creation events for the containers we were asked to
